Split Encryption in 2
